+++ src/main/java/step3/InputView.java
@@ -1,4 +1,4 @@
-package step2;
+package step3;
 
 import java.util.Scanner;
