+++ src/main/java/step3/LottoFactory.java
@@ -1,21 +1,19 @@
-package step2;
+package step3;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
 
-public class Lotto {
+public class LottoFactory {
 	private final int LOTTO_START = 1;
 	private final int LOTTO_END = 45;
 	private final int LOTTO_PICK_START = 0;
 	private final int LOTTO_PICK_END = 5;
 	private List<Integer> balls = new ArrayList<>();
 
-	public Lotto() {
+	public LottoFactory() {
 		for (int i = LOTTO_START; i <= LOTTO_END ; i++) {
 			this.balls.add(i);
 		}
