+++ src/main/java/step3/Main.java
@@ -1,4 +1,4 @@
-package step2;
+package step3;
 
 public class Main {
 	public static void main(String[] args) {
