+++ src/main/java/step3/Operator.java
@@ -1,4 +1,4 @@
-package step2;
+package step3;
 
 import java.util.Collections;
 import java.util.Map;
