+++ src/test/java/step3/LottoTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step3;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,14 +8,14 @@ import org.junit.jupiter.api.Test;
 public class LottoTest {
 	@Test
 	void create() {
-		Lotto lotto = new Lotto();
+		LottoFactory lotto = new LottoFactory();
 		assertThat(lotto.getBalls().size()).isEqualTo(45);
 	}
 
 	@Test
 	@DisplayName("자동 추첨 결과")
 	void makeResult() {
-		Lotto lotto = new Lotto();
+		LottoFactory lotto = new LottoFactory();
 		assertThat(lotto.getRandomLotto().size()).isEqualTo(6);
 	}
 }
